refactor: deploy xvs separately
